refactor: auth controllers
