+++ waltid-services/waltid-wallet-api/src/main/kotlin/id/walt/webwallet/Main.kt
@@ -13,9 +13,10 @@ import id.walt.webwallet.web.Administration.configureAdministration
 import id.walt.webwallet.web.controllers.*
 import id.walt.webwallet.web.controllers.NotificationController.notifications
 import id.walt.webwallet.web.controllers.PushController.push
-import id.walt.webwallet.web.controllers.auth.auth
+import id.walt.webwallet.web.controllers.auth.defaultAuthRoutes
+import id.walt.webwallet.web.controllers.auth.keycloak.keycloakAuthRoutes
+import id.walt.webwallet.web.controllers.auth.oidc.oidcAuthRoutes
 import id.walt.webwallet.web.plugins.*
-import id.walt.webwallet.web.plugins.walletOpenApiPluginAmendment
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.server.application.*
 import org.bouncycastle.jce.provider.BouncyCastleProvider
@@ -59,7 +60,9 @@ fun Application.webWalletModule(withPlugins: Boolean = true) {
     if (withPlugins) {
         configurePlugins()
     }
-    auth()
+    { oidcAuthRoutes() } whenFeature FeatureCatalog.oidcAuthenticationFeature
+    defaultAuthRoutes()
+    keycloakAuthRoutes()
     accounts();
     { push() } whenFeature FeatureCatalog.pushFeature
