+++ waltid-services/waltid-wallet-api/src/main/kotlin/id/walt/webwallet/web/controllers/auth/keycloak/KeycloakRegisterController.kt
@@ -1,0 +1,30 @@
+package id.walt.webwallet.web.controllers.auth.keycloak
+
+import id.walt.webwallet.web.controllers.auth.RegisterControllerBase
+import id.walt.webwallet.web.model.AccountRequest
+import id.walt.webwallet.web.model.KeycloakAccountRequest
+import io.github.smiley4.ktorswaggerui.dsl.routes.OpenApiRoute
+import io.ktor.http.*
+
+class KeycloakRegisterController : RegisterControllerBase(keycloakAuthPath, keycloakAuthTags) {
+    override fun apiBuilder(): OpenApiRoute.() -> Unit = {
+        summary = "Keycloak registration with [username + email + password]"
+        description = "Creates a user in the configured Keycloak instance."
+        request {
+            body<AccountRequest> {
+                example("username + email + password") {
+                    value = KeycloakAccountRequest(
+                        username = "Max_Mustermann",
+                        email = "user@email.com",
+                        password = "password",
+                        token = "eyJhb..."
+                    )
+                }
+            }
+        }
+        response {
+            HttpStatusCode.Created to { description = "Registration succeeded " }
+            HttpStatusCode.BadRequest to { description = "Registration failed" }
+        }
+    }
+}
